Pins debugger HAL
